:bento: Chore: 그래프 api data class 정의

+++ app/src/main/java/com/example/snack4diet/api/analysisGraph/Data.kt
@@ -9,5 +9,5 @@ data class Data(
     val fatLastSevenDays: List<Int>,
     val proteinLastFourWeek: List<Int>,
     val proteinLastSevenDays: List<Int>,
-    val totalScore: Int
+    val totalScore: Double
 )
